+++ package/src/components/PopoutRootView.tsx
@@ -305,7 +305,7 @@ const PopoutRootView = ({ children }: { children: ReactNode }) => {
             item={elementOpened}
             panScale={panScale}
             backdropProgress={backdropProgress}
-            disableBlur={!overlayConfig.overlayBlur}
+            disableBlur={overlayConfig.overlayBlur === false}
             overlayImageStyle={overlayConfig.overlayImageStyle}
           >
             {typeof OverlayComponent === 'function' ? (