+++ package/CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+- Possible fix for disabling overlayBlur
+- Fixed closebutton position
+
 ## [0.1.15-8] - 2025-02-02
 
 - overlayNotchInset was reversed. Fixed it.